requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
spacy>=3.6.0
nltk>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pyyaml>=6.0

